fix: represent MB as float64

+++ internal/domain/output_video.go
@@ -3,8 +3,8 @@ package domain
 import "path/filepath"
 
 type OutputVideo struct {
-	TargetSize            int          `json:"targetSize"`            // bytes
-	TargetSizeAfterMargin int          `json:"targetSizeAfterMargin"` // bytes
+	TargetSize            float64      `json:"targetSize"`            // MB
+	TargetSizeAfterMargin float64      `json:"targetSizeAfterMargin"` // MB
 	OutputPath            string       `json:"outputPath"`
 	OutputName            string       `json:"outputName"`
 	Size                  int          `json:"size"`        // bytes
